requirements.txt
# AI Workload Benchmarking Framework - Dependencies
# Core inference backends
openvino>=2023.3.0
openvino-genai>=2024.0.0
onnxruntime>=1.16.0
huggingface_hub>=0.23.0

# Web dashboard
streamlit>=1.28.0

# Data and analytics
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.40.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Image handling
opencv-python-headless>=4.8.0

# System monitoring & stress testing
psutil>=5.9.0
py-cpuinfo>=9.0.0
GPUtil>=1.4.0

# Utilities
python-dotenv>=1.0.0
Pillow>=10.0.0

# Export
openpyxl>=3.1.0
